+++ src/styles/global.style.css
@@ -78,46 +78,3 @@ select:-webkit-autofill:focus  */
   pointer-events: all;
   @apply range-slider-thumb-base;
 }
-
-/* *Google Map Custom Style */
-.gm-style {
-  .gm-style-iw-c {
-    padding-inline-end: 12px !important;
-    padding-bottom: 12px !important;
-    background-color: var(--tertiary-inverted) !important;
-    backdrop-filter: blur(2px);
-  }
-  .gm-style-iw-tc::after {
-    background: var(--tertiary-inverted) !important;
-    backdrop-filter: blur(2px);
-  }
-  .gm-style-iw-ch {
-    padding-top: 0px !important;
-    & > .custom-logo {
-      width: 48px;
-      height: 48px;
-      object-fit: contain;
-      display: block;
-    }
-  }
-  .gm-style-iw-chr {
-    display: flex;
-    align-items: center;
-    & > .gm-ui-hover-effect {
-      opacity: 1;
-      display: flex !important;
-      align-items: center;
-      justify-content: center;
-      width: 24px !important;
-      height: 24px !important;
-      padding: 0 !important;
-      & > span {
-        display: none !important;
-      }
-    }
-    .custom-close-icon {
-      width: 24px !important;
-      height: 24px !important;
-    }
-  }
-}
